musiki_backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
librosa==0.10.2
scipy==1.14.1
numpy==1.26.4
gunicorn==23.0.0
Pillow==11.1.0
